hls: Support to compile CIRCT and added hls-test-suit
New make targets for compiling CIRCT and associated LLVM and MLIR
New hls-test-suit with initial C tests

The build installs circt, llvm, and mlir in circt/local.
This provides a single location for bin, include and lib.
It also makes it possible to delete the build directories to save
disk space, with 'make circt-clean'.

Added all the tests from the Regression_test directory of the
dynamatic repository: github.com/lana555/dynamatic.
The tests have been modified to print the result to stdout such
that the execution of the tests can be verified.